test(BDD스타일로 변경)
