+++ src/test/java/kitchenpos/bo/MockTest.java
@@ -1,0 +1,4 @@
+package kitchenpos.bo;
+
+public class MockTest {
+}
